+++ src/main/java/uk/gov/hmcts/reform/dg/docassembly/service/TemplateRenditionService.java
@@ -1,46 +1,29 @@
 package uk.gov.hmcts.reform.dg.docassembly.service;
 
-import okhttp3.MultipartBody;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
 import okhttp3.Response;
 import org.apache.tomcat.util.http.fileupload.IOUtils;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.dg.docassembly.appinsights.DependencyProfiler;
 import uk.gov.hmcts.reform.dg.docassembly.dto.CreateTemplateRenditionDto;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.Base64;
-import java.util.UUID;
 
 @Service
 public class TemplateRenditionService {
 
-    private final String docmosisUrl;
-
-    private final String docmosisAccessKey;
-
-    private final OkHttpClient httpClient;
-
     private final DmStoreUploader dmStoreUploader;
+    private final DocmosisApiClient docmosisApiClient;
 
-    public TemplateRenditionService(OkHttpClient httpClient,
-                                    DmStoreUploader dmStoreUploader,
-                                    @Value("${docmosis.convert.endpoint}") String docmosisUrl,
-                                    @Value("${docmosis.accessKey}") String docmosisAccessKey) {
-        this.httpClient = httpClient;
+    public TemplateRenditionService(DmStoreUploader dmStoreUploader, DocmosisApiClient docmosisApiClient) {
         this.dmStoreUploader = dmStoreUploader;
-        this.docmosisUrl = docmosisUrl;
-        this.docmosisAccessKey = docmosisAccessKey;
+        this.docmosisApiClient = docmosisApiClient;
     }
 
     public CreateTemplateRenditionDto renderTemplate(CreateTemplateRenditionDto createTemplateRenditionDto)
             throws IOException {
 
-        Response response = this.render(createTemplateRenditionDto);
+        Response response = this.docmosisApiClient.render(createTemplateRenditionDto);
 
         if (!response.isSuccessful()) {
             throw new TemplateRenditionException(
@@ -57,36 +40,5 @@ public class TemplateRenditionService {
         dmStoreUploader.uploadFile(file, createTemplateRenditionDto);
 
         return createTemplateRenditionDto;
-    }
-
-    @DependencyProfiler(name = "docmosis", action = "render")
-    public Response render(CreateTemplateRenditionDto createTemplateRenditionDto) throws IOException {
-        String tempFileName = String.format("%s%s",
-                UUID.randomUUID().toString(),
-                createTemplateRenditionDto.getOutputType().getFileExtension());
-
-        MultipartBody requestBody = new MultipartBody
-                .Builder()
-                .setType(MultipartBody.FORM)
-                .addFormDataPart(
-                        "templateName",
-                        new String(Base64.getDecoder().decode(createTemplateRenditionDto.getTemplateId())))
-                .addFormDataPart(
-                        "accessKey",
-                        docmosisAccessKey)
-                .addFormDataPart(
-                        "outputName",
-                        tempFileName)
-                .addFormDataPart(
-                        "data",
-                        String.valueOf(createTemplateRenditionDto.getFormPayload()))
-                .build();
-
-        Request request = new Request.Builder()
-                .url(docmosisUrl)
-                .method("POST", requestBody)
-                .build();
-
-        return httpClient.newCall(request).execute();
     }
 }
