+++ src/test/java/uk/gov/hmcts/reform/dg/docassembly/service/TemplateRenditionServiceTest.java
@@ -38,10 +38,8 @@ public class TemplateRenditionServiceTest {
         dmStoreUploader = Mockito.mock(DmStoreUploader.class);
 
         templateRenditionService = new TemplateRenditionService(
-                client,
-                dmStoreUploader,
-                "http://tonrnado.com",
-                "x"
+            dmStoreUploader,
+            new DocmosisApiClient(client, "http://tornado.com", "x")
         );
 
         createTemplateRenditionDto = createTemplateRenditionDto();
